+++ CHANGELOG.md
@@ -69,6 +69,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - `[s]` will strike through text
 - `AsyncOptionsView` now supports configuring its fade duration.
 - `MarkupPalette` and the associated `PaletteMarkerProcessor` now support custom TMP rich text tags beyond what the more basic markers allow.
+- `LineMetadata` now has public API methods for constructing and manually adding elements.
+- `DialogueRunner.SetProject` now sets the `Program` of its internal `Dialogue` object. Previously, this didn't happen until `StartDialogue()`.
 - Added a sample (Advanced Saliency) making use of templated nodes and built in saliency to show off creating storylet vignettes.
 
 ### Changed
@@ -90,6 +92,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fixed an issue where the System.Threading.Tasks implementation of YarnTask.WaitUntil did not return early when the CancellationToken was cancelled ([@dogboydog](https://github.com/YarnSpinnerTool/YarnSpinner-Unity/pull/304))
 - If a `VoiceOverView` is configured to automatically advance at the end of audio playback, when a line that has no `AudioClip` is delivered, the `VoiceOverView` will now advance the line immediately after logging an error.
 - Renamed `FormatForMarker` to `PaletteForMarker` in `MarkupPalette`.
+- Generated variable storage code now represents string values as signed integers, not unsigned integers. (Unity doesn't correctly serialize enum values over the signed integer maximum.)
+- `YarnNode`-attributed variables can now be configured to allow editing as a text field even when the Yarn Project is not set. This is now the default behaviour for `DialogueReference`.
 
 ### Removed